Moves the BlazeManager structure into its own file
