+++ src/apps/chifra/internal/scrape/handle_blaze_blaze.go
@@ -1,7 +1,6 @@
 package scrapePkg
 
 import (
-	"encoding/json"
 	"fmt"
 	"os"
 	"sort"
@@ -15,47 +14,8 @@ import (
 	"github.com/TrueBlocks/trueblocks-core/src/apps/chifra/pkg/logger"
 	"github.com/TrueBlocks/trueblocks-core/src/apps/chifra/pkg/rpc"
 	"github.com/TrueBlocks/trueblocks-core/src/apps/chifra/pkg/tslib"
-	"github.com/TrueBlocks/trueblocks-core/src/apps/chifra/pkg/types"
 	"github.com/TrueBlocks/trueblocks-core/src/apps/chifra/pkg/utils"
 )
-
-// scrapedData combines the block data, trace data, and log data into a single structure
-type scrapedData struct {
-	blockNumber base.Blknum
-	traces      []types.SimpleTrace
-	logs        []types.SimpleLog
-}
-
-type BlazeManager struct {
-	Chain        string                  `json:"chain"`
-	NChannels    uint64                  `json:"nChannels"`
-	StartBlock   uint64                  `json:"startBlock"`
-	BlockCount   uint64                  `json:"blockCnt"`
-	UnripeDist   uint64                  `json:"unripe"`
-	RpcProvider  string                  `json:"rpcProvider"`
-	AppsPerChunk uint64                  `json:"-"`
-	RipeBlock    uint64                  `json:"ripeBlock"`
-	NProcessed   uint64                  `json:"nProcessed"`
-	Timestamps   []tslib.TimestampRecord `json:"-"`
-	ProcessedMap map[base.Blknum]bool    `json:"-"`
-	BlockWg      sync.WaitGroup          `json:"-"`
-	AppearanceWg sync.WaitGroup          `json:"-"`
-	TimestampsWg sync.WaitGroup          `json:"-"`
-}
-
-func (blazeMan *BlazeManager) String() string {
-	copy := BlazeManager{
-		Chain:      blazeMan.Chain,
-		NChannels:  blazeMan.NChannels,
-		StartBlock: blazeMan.StartBlock,
-		BlockCount: blazeMan.BlockCount,
-		UnripeDist: blazeMan.UnripeDist,
-		RipeBlock:  blazeMan.RipeBlock,
-		NProcessed: blazeMan.NProcessed,
-	}
-	b, _ := json.MarshalIndent(&copy, "", "  ")
-	return string(b)
-}
 
 // HandleBlaze does the actual scraping, walking through block_cnt blocks and querying traces and logs
 // and then extracting addresses and timestamps from those data structures.
